"origin/RDCC-529, adding to unit test - still failing, but progress made"
