+++ src/main/java/uk/gov/hmcts/reform/professionalapi/service/PrdEnumType.java
@@ -1,0 +1,5 @@
+package uk.gov.hmcts.reform.professionalapi.service;
+
+public enum PrdEnumType {
+    SIDAM_ROLE, ADMIN_ROLE, JURISD_ID
+}
